requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-decouple==3.8
django-import-export==3.3.6
openpyxl==3.1.2
reportlab==4.1.0
drf-yasg==1.21.7
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.1
